Restructured and started on wolf cluster
Basically implemented a simple fifo queue specific
for this problem. I've tested that it works properly.
Now for the implementation of the cluster method there
seem to be some extra stuff needed in theory to figure
out what the acceptance ratio should be. I'll read about it
tomorrow i guess...